+++ CHANGELOG.md
@@ -6,9 +6,12 @@
 * Added `onMessage()` method to HAMqtt class
 * Added support for HA Covers
 * Added support for setting different prefix for non-discovery topics (see [Advanced MQTT example](examples/mqtt-advanced/mqtt-advanced.ino))
+* Added `setName` method to HASensor
+* Added `setName` method to HASwitch
 
 **Updates:**
 * Removed legacy properties from HAFan (Home Assistant 2021.4.4). Deprecated methods will be removed after a quarter (2021.7)
+* Separated `uniqueID` field from `name` in all devices types
 
 ## 1.2.0